Revert "v1.5.0 (#7)"
This reverts commit e9b324d2313b92a53e1ade5f5657d8cf4de5dc48.
