+++ lastfm/love.py
@@ -2,8 +2,9 @@ import discord
 from redbot.core.utils.menus import DEFAULT_CONTROLS, menu
 
 from .abc import MixinMeta
-from .exceptions import *
-from .fmmixin import command_fm
+from .errors import *
+from .fmmixin import fm
+from .utils import *
 
 
 class LoveMixin(MixinMeta):
@@ -13,6 +14,7 @@ class LoveMixin(MixinMeta):
         params = {
             "api_key": self.token,
             "artist": artist,
+            "method": "track.unlove",
             "sk": key,
             "track": track,
         }
@@ -23,8 +25,8 @@ class LoveMixin(MixinMeta):
         data = await self.api_post(params=params)
         return data
 
-    @command_fm.command(name="love", usage="<track name> | <artist name>")
-    async def command_love(self, ctx, *, track=None):
+    @fm.command(usage="<track name> | <artist name>")
+    async def love(self, ctx, *, track=None):
         """
         Love a song on last.fm.
 
@@ -33,7 +35,7 @@ class LoveMixin(MixinMeta):
             [p]love <track name> | <artist name>
         """
         conf = await self.config.user(ctx.author).all()
-        self.check_if_logged_in_and_sk(conf)
+        check_if_logged_in_and_sk(conf)
         if track:
             try:
                 trackname, artistname = [x.strip() for x in track.split("|")]
@@ -42,9 +44,19 @@ class LoveMixin(MixinMeta):
             except ValueError:
                 return await ctx.send("\N{WARNING SIGN} Incorrect format! use `track | artist`")
         else:
-            trackname, artistname, albumname, imageurl = await self.get_current_track(
-                ctx, conf["lastfm_username"]
+            data = await self.api_request(
+                ctx,
+                {
+                    "user": conf["lastfm_username"],
+                    "method": "user.getrecenttracks",
+                    "limit": 1,
+                },
             )
+            tracks = data["recenttracks"]["track"]
+            if not tracks:
+                return await ctx.send("You have not listened to anything yet!")
+            artistname = tracks[0]["artist"]["#text"]
+            trackname = tracks[0]["name"]
 
         data = await self.api_request(
             ctx,
@@ -64,11 +76,11 @@ class LoveMixin(MixinMeta):
         result = await self.love_or_unlove_song(
             data["track"]["name"], data["track"]["artist"]["name"], True, conf["session_key"]
         )
-        await self.maybe_send_403_msg(ctx, result)
+        await maybe_send_403_msg(self, ctx, result)
         await ctx.send(f"Loved **{trackname[:50]}** by **{artistname[:50]}**")
 
-    @command_fm.command(name="unlove", usage="<track name> | <artist name>")
-    async def command_unlove(self, ctx, *, track=None):
+    @fm.command(usage="<track name> | <artist name>")
+    async def unlove(self, ctx, *, track=None):
         """
         Unlove a song on last.fm.
 
@@ -77,7 +89,7 @@ class LoveMixin(MixinMeta):
             [p]unlove <track name> | <artist name>
         """
         conf = await self.config.user(ctx.author).all()
-        self.check_if_logged_in_and_sk(conf)
+        check_if_logged_in_and_sk(conf)
         if track:
             try:
                 trackname, artistname = [x.strip() for x in track.split("|")]
@@ -86,9 +98,19 @@ class LoveMixin(MixinMeta):
             except ValueError:
                 return await ctx.send("\N{WARNING SIGN} Incorrect format! use `track | artist`")
         else:
-            trackname, artistname, albumname, imageurl = await self.get_current_track(
-                ctx, conf["lastfm_username"]
+            data = await self.api_request(
+                ctx,
+                {
+                    "user": conf["lastfm_username"],
+                    "method": "user.getrecenttracks",
+                    "limit": 1,
+                },
             )
+            tracks = data["recenttracks"]["track"]
+            if not tracks:
+                return await ctx.send("You have not listened to anything yet!")
+            artistname = tracks[0]["artist"]["#text"]
+            trackname = tracks[0]["name"]
 
         data = await self.api_request(
             ctx,
@@ -108,11 +130,11 @@ class LoveMixin(MixinMeta):
         result = await self.love_or_unlove_song(
             data["track"]["name"], data["track"]["artist"]["name"], False, conf["session_key"]
         )
-        await self.maybe_send_403_msg(ctx, result)
+        await maybe_send_403_msg(self, ctx, result)
         await ctx.send(f"Unloved **{trackname[:50]}** by **{artistname[:50]}**")
 
-    @command_fm.command(name="loved")
-    async def command_loved(self, ctx, user: discord.User = None):
+    @fm.command()
+    async def loved(self, ctx, user: discord.User = None):
         """
         Get a list of loved songs for a user.
 
@@ -123,7 +145,7 @@ class LoveMixin(MixinMeta):
         if not user:
             user = ctx.author
         conf = await self.config.user(user).all()
-        self.check_if_logged_in_and_sk(conf)
+        check_if_logged_in_and_sk(conf)
         data = await self.api_request(
             ctx, {"user": conf["lastfm_username"], "method": "user.getlovedtracks"}
         )
@@ -133,7 +155,7 @@ class LoveMixin(MixinMeta):
         tracks = [f"{x['name']} by {x['artist']['name']}\n" for x in tracks]
         content = discord.Embed(color=await ctx.embed_color(), title=f"{user.name}'s loved tracks")
 
-        pages = await self.create_pages(content, tracks)
+        pages = await create_pages(content, tracks)
         for i, page in enumerate(pages):
             page.set_footer(text=f"Page {i + 1}/{len(pages)}")
         if len(pages) > 1:
