+++ lastfm/top.py
@@ -3,20 +3,21 @@ from redbot.core.utils.chat_formatting import escape
 from redbot.core.utils.menus import DEFAULT_CONTROLS, menu
 
 from .abc import MixinMeta
-from .exceptions import *
-from .fmmixin import command_fm
+from .errors import *
+from .fmmixin import fm
+from .utils import *
 
 
 class TopMixin(MixinMeta):
     """Top Artist/Album/Track Commands"""
 
-    @command_fm.command(name="topartists", aliases=["ta"], usage="[timeframe] [amount]")
-    async def command_topartists(self, ctx, *args):
+    @fm.command(aliases=["ta"], usage="[timeframe] [amount]")
+    async def topartists(self, ctx, *args):
         """Most listened artists."""
         conf = await self.config.user(ctx.author).all()
-        self.check_if_logged_in(conf)
+        check_if_logged_in(conf)
         async with ctx.typing():
-            arguments = self.parse_arguments(args)
+            arguments = parse_arguments(args)
             data = await self.api_request(
                 ctx,
                 {
@@ -36,30 +37,32 @@ class TopMixin(MixinMeta):
             for i, artist in enumerate(artists, start=1):
                 name = escape(artist["name"], formatting=True)
                 plays = artist["playcount"]
-                rows.append(f"`#{i:2}` **{plays}** {self.format_plays(plays)} — **{name}**")
+                rows.append(f"`#{i:2}` **{plays}** {format_plays(plays)} — **{name}**")
 
             image_url = await self.scrape_artist_image(artists[0]["name"], ctx)
+            # image_colour = await color_from_image_url(image_url)
 
             content = discord.Embed(color=await self.bot.get_embed_color(ctx.channel))
+            # content.colour = int(image_colour, 16)
             content.set_thumbnail(url=image_url)
             content.set_footer(text=f"Total unique artists: {user_attr['total']}")
             content.set_author(
-                name=f"{user_attr['user']} — {self.humanized_period(arguments['period']).capitalize()} top artists",
+                name=f"{user_attr['user']} — {humanized_period(arguments['period']).capitalize()} top artists",
                 icon_url=ctx.message.author.avatar_url,
             )
 
-        pages = await self.create_pages(content, rows)
+        pages = await create_pages(content, rows)
         if len(pages) > 1:
             await menu(ctx, pages[:15], DEFAULT_CONTROLS)
         else:
             await ctx.send(embed=pages[0])
 
-    @command_fm.command(name="topalbums", aliases=["talb"], usage="[timeframe] [amount]")
-    async def command_topalbums(self, ctx, *args):
+    @fm.command(aliases=["talb"], usage="[timeframe] [amount]")
+    async def topalbums(self, ctx, *args):
         """Most listened albums."""
         conf = await self.config.user(ctx.author).all()
-        self.check_if_logged_in(conf)
-        arguments = self.parse_arguments(args)
+        check_if_logged_in(conf)
+        arguments = parse_arguments(args)
         data = await self.api_request(
             ctx,
             {
@@ -81,32 +84,35 @@ class TopMixin(MixinMeta):
             artist_name = escape(album["artist"]["name"], formatting=True)
             plays = album["playcount"]
             rows.append(
-                f"`#{i:2}` **{plays}** {self.format_plays(plays)} — **{artist_name}** — ***{name}***"
+                f"`#{i:2}` **{plays}** {format_plays(plays)} — **{artist_name}** — ***{name}***"
             )
 
         image_url = albums[0]["image"][-1]["#text"]
+        # image_url_small = albums[0]['image'][1]['#text']
+        # image_colour = await color_from_image_url(image_url_small)
 
         content = discord.Embed(color=await self.bot.get_embed_color(ctx.channel))
+        # content.colour = int(image_colour, 16)
         content.set_thumbnail(url=image_url)
         content.set_footer(text=f"Total unique albums: {user_attr['total']}")
         content.set_author(
-            name=f"{user_attr['user']} — {self.humanized_period(arguments['period']).capitalize()} top albums",
+            name=f"{user_attr['user']} — {humanized_period(arguments['period']).capitalize()} top albums",
             icon_url=ctx.message.author.avatar_url,
         )
 
-        pages = await self.create_pages(content, rows)
+        pages = await create_pages(content, rows)
         if len(pages) > 1:
             await menu(ctx, pages[:15], DEFAULT_CONTROLS)
         else:
             await ctx.send(embed=pages[0])
 
-    @command_fm.command(name="toptracks", aliases=["tt"], usage="[timeframe] [amount]")
-    async def command_toptracks(self, ctx, *args):
+    @fm.command(aliases=["tt"], usage="[timeframe] [amount]")
+    async def toptracks(self, ctx, *args):
         """Most listened tracks."""
         conf = await self.config.user(ctx.author).all()
-        self.check_if_logged_in(conf)
+        check_if_logged_in(conf)
         async with ctx.typing():
-            arguments = self.parse_arguments(args)
+            arguments = parse_arguments(args)
             data = await self.api_request(
                 ctx,
                 {
@@ -128,7 +134,7 @@ class TopMixin(MixinMeta):
                 artist_name = escape(track["artist"]["name"], formatting=True)
                 plays = track["playcount"]
                 rows.append(
-                    f"`#{i:2}` **{plays}** {self.format_plays(plays)} — **{artist_name}** — ***{name}***"
+                    f"`#{i:2}` **{plays}** {format_plays(plays)} — **{artist_name}** — ***{name}***"
                 )
             trackdata = await self.api_request(
                 ctx,
@@ -154,11 +160,11 @@ class TopMixin(MixinMeta):
 
             content.set_footer(text=f"Total unique tracks: {user_attr['total']}")
             content.set_author(
-                name=f"{user_attr['user']} — {self.humanized_period(arguments['period']).capitalize()} top tracks",
+                name=f"{user_attr['user']} — {humanized_period(arguments['period']).capitalize()} top tracks",
                 icon_url=ctx.message.author.avatar_url,
             )
 
-            pages = await self.create_pages(content, rows)
+            pages = await create_pages(content, rows)
             if len(pages) > 1:
                 await menu(ctx, pages[:15], DEFAULT_CONTROLS)
             else:
